pyproject.toml
[project]
name = "loan-agent-workflow"
version = "0.1.0"
description = "Multi-agent loan processing system using Microsoft Agent Framework"
authors = [
    { name = "Microsoft Agent Framework Team", email = "agent-framework@microsoft.com" }
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Core dependencies - always required
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "httpx>=0.27.0",
    "rich>=13.0.0", # For better console output
    "mcp[cli]>=1.12.3", # Model Context Protocol for agent tools
    "uvicorn>=0.35.0",
    "requests>=2.32.4",
    "PyYAML>=6.0.0", # For configuration file parsing
    # Logging infrastructure (always available)
    "structlog>=24.1.0", # Structured logging with processors
    "opentelemetry-api>=1.20.0", # OTEL API for correlation tracking
    # Microsoft Agent Framework (to be added when available)
    # "microsoft-agent-framework>=1.0.0",
]

[project.optional-dependencies]
# Future AI provider options
openai = [
    "openai>=1.0.0",
]
azure = [
    "azure-openai>=1.0.0",
    "azure-monitor-opentelemetry>=1.6.13", # Azure Application Insights integration
    "opentelemetry-instrumentation-logging", # Auto trace context injection
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pytest-cov>=6.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
extend-select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["loan_processing"]
known-third-party = ["agents"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 120
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.hatch.build.targets.wheel]
packages = ["loan_processing"]

[tool.hatch.build.targets.sdist]
include = [
  "loan_processing",
  "README.md",
  "LICENSE",
  "agent-persona",
  "docs",
  "pyproject.toml",
]
